+++ src/Effects/WorldEffects/SpawnCritters.cs
@@ -41,6 +41,12 @@ public sealed class SpawnCritters : CrowdControlEffect
         NPCID.GoldBunny, NPCID.GoldButterfly, NPCID.GoldDragonfly, NPCID.GoldFrog, NPCID.GoldGoldfish, NPCID.GoldGrasshopper, NPCID.GoldLadyBug,
         NPCID.GoldMouse, NPCID.GoldSeahorse, NPCID.SquirrelGold, NPCID.GoldWaterStrider, NPCID.GoldWorm, NPCID.BunnySlimed, NPCID.BunnyXmas,
         NPCID.PartyBunny, NPCID.Dolphin, NPCID.SeaTurtle, NPCID.CrimsonBunny, NPCID.CorruptBunny
+
+        // 1.4.4
+        /*
+        NPCID.GrayCockatiel, NPCID.YellowCockatiel, NPCID.Toucan, NPCID.BlueMacaw, NPCID.ScarletMacaw, 
+        NPCID.Stinkbug, NPCID.Shimmerfly, BoundTownSlimeYellow
+        */
     };
 
     private static readonly string[] CalamityCritters =