+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.8] - 2022-12-11
+
 - Added "Spawn Random Town NPC" effect.
 - Added "Provide Swimming Buffs" effect.
 - Added "Shuffle Sfx" effect, inspired by Hollow Knight Crowd Control.